+++ CHANGELOG.md
@@ -8,9 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- The next version number should be the version already set in pubspec.yaml -->
 ### Fixed
 - Android: Fix various compilation warnings [#634](https://github.com/rokwire/app-flutter-plugin/issues/634).
+- Survey result rule evaluation issues [#637](https://github.com/rokwire/app-flutter-plugin/issues/637).
 ### Added
 - Acknowledge Reaction innerContext field [#5238](https://github.com/rokwire/illinois-app/issues/5238).
 - Added ListUtils.last() helper [#5344](https://github.com/rokwire/illinois-app/issues/5344).
+- Added ListUtils.stripNull() helper [#5344](https://github.com/rokwire/illinois-app/issues/5344).
+### Changed
+- ImageUtils.mapGroupMarkerImage extended to handle pin markers [#5344](https://github.com/rokwire/illinois-app/issues/5344).
 
 ## [1.11.0] - 2025-08-05
 ### Changed